:recycle: refactor: flowbuilder and packages map parser
