+++ flowrs-build-designer/rete/index.ts
@@ -1,13 +1,13 @@
-import { createEditor as createDefaultEditor } from './default'
-import { createEditor as createFlowBuilder } from './flowbuilder'
+import { createEditor as createExampleEditor } from './example'
+import { createEditor as createFlowrsEditor } from './flowrs/editor'
 
 const factory = {
-  'flowbuilder': createFlowBuilder,
-  'default': createDefaultEditor
+  'flowbuilder': createFlowrsEditor,
+  'example': createExampleEditor
 }
 // eslint-disable-next-line no-restricted-globals, no-undef
 const query = typeof location !== 'undefined' && new URLSearchParams(location.search)
-const name = ((query && query.get('template')) || 'flowbuilder') as keyof typeof factory
+const name = (query && query.get('template') || 'flowbuilder') as keyof typeof factory
 
 const createEditor = factory[name]
